Fixup! Continue refactoring
