+++ src/client/components/PersonalisedDashboard/state.js
@@ -1,3 +1,3 @@
 export const TASK_USER_DETAILS = 'TASK_USER_DETAILS'
-export const ID = 'getUserDetails'
+export const ID = 'userDetails'
 export const state2props = (state) => state[ID]
